Get tests working and compiling after extracting morton helpers
