+++ CMakeLists.txt
@@ -55,6 +55,7 @@ endforeach()
 add_custom_target(compile_shaders ALL DEPENDS ${SHADER_SPV_FILES})
 
 add_executable(tests
+    ${TEST_SRC_DIR}/test_morton_helpers.cpp 
     ${TEST_SRC_DIR}/test_morton.cpp
     ${TEST_SRC_DIR}/test_mynydd.cpp
     ${TEST_SRC_DIR}/test_shader_kernels.cpp
